+++ test/index.js
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import * as Pinpoint from '../lib/index.js';
+import * as Pinpoint from '../src/index.mjs';
 
 describe('Pinpoint', () => {
     describe('location()', () => {
